backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-git==2020.9.0
docker==7.1.0
psutil==6.0.0
firebase-admin==6.5.0
pydantic==2.9.2
aiohttp==3.10.5